+++ src/__tests__/SearchBar.test.js
@@ -3,7 +3,7 @@ import { Provider } from "react-redux";
 import { render, fireEvent, act, waitFor } from "@testing-library/react";
 import configureStore from "redux-mock-store";
 import { addCity } from "../state/actions";
-import SearchBar from "../components/pages/Nav/SearchBar";
+import { SearchBar } from "../components/pages/SearchBar";
 const mockStore = configureStore([]);
 describe("<SearchBar />", () => {
   let store, component;
