+++ src/components/pages/SearchBar/index.js
@@ -1,0 +1,1 @@
+export { default as SearchBar } from "./SearchBarContainer";
